An S4 class for the Laplace mechanism of differential privacy.

A class that implements the basic Laplace mechanism of differential privacy, for privatizing numeric vector releases.

# S4 method for DPMechLaplace
show(object)

Arguments

object
an instance of class DPMech.

Methods (by generic)

show: automatically prints the object.

Slots

sensitivity

non-negative scalar numeric L1 target sensitivity. Defaults to Inf for use with sensitivitySampler().

target

the target non-private function to be privatized, takes lists. Defaults to a constant function. Laplace mechanism assumes functions that release numeric vectors of fixed dimension dims.

gammaSensitivity

NA_real_ if inactive, or scalar in [0,1) indicating that responses must be RDP with specific confidence.

dims

positive scalar numeric dimension of responses. Defaults to NA_integer_ for use with sensitivitySampler() which can probe target to determine dimension.

References

Cynthia Dwork, Frank McSherry, Kobbi Nissim, and Adam Smith. "Calibrating noise to sensitivity in private data analysis." In Theory of Cryptography Conference, pp. 265-284. Springer Berlin Heidelberg, 2006.